What's your favorite song?
Think of all the attributes that you could use to describe that song.
That is: all of it's details or "meta-data". These are attributes like
"Artist", "Year Released", "Genre", "Duration", etc. Think of as many
different characteristics as you can.

Create a new project and within your main file list all of the
attributes of the song, one after another. Create a variable for each
attribute, and give each variable a value. Here's an example:

```
String genre = "Jazz";

```

Write each variable on its own line. Then, after you have listed the
variables, print each one of them out. Think of as many characteristics
of the song as you can. Try to use all of the data types you have
learned so far.